MCTS agent moved to its own folder
