+++ Stratega/include/Stratega/Agent/TreeSearchAgents/ActionAbstractionMCTSAgent.h
@@ -1,11 +1,11 @@
 #pragma once
 #include <Stratega/Agent/Agent.h>
-#include <Stratega/Agent/TreeSearchAgents/MCTSNode.h>
+#include <Stratega/Agent/MCTSAgent/MCTSNode.h>
 #include <Stratega/Agent/Heuristic/AbstractHeuristic.h>
 #include <Stratega/Agent/ActionScripts/AttackClosestOpponentScript.h>
 #include <Stratega/Agent/ActionScripts/AttackWeakestOpponentScript.h>
 
-#include "MCTSParameters.h"
+#include "Stratega/Agent/MCTSAgent/MCTSParameters.h"
 
 namespace SGA
 {
